backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.53
azure-ai-vision-imageanalysis==1.0.0b1
azure-identity==1.15.0
openai==1.6.1
python-dotenv==1.0.0
cryptography>=41.0.0
bcrypt==4.1.2
flask-cors==4.0.0
pytest==7.4.3
hypothesis==6.92.1
Pillow==10.1.0
pyotp==2.9.0
qrcode[pil]==8.2
Werkzeug==2.3.7